Fix virtual currency reward amount type to integer.
Align reward payload modeling with backend semantics by decoding `amount` as Int and updating tests to reject fractional values as malformed rewards.

+++ Sources/Ads/RewardVerification/Networking/RewardVerificationStatusResponse.swift
@@ -138,7 +138,7 @@ extension RewardVerificationStatusResponse: Decodable {
         switch rewardType {
         case RewardType.virtualCurrency:
             guard let code = try? rewardContainer.decode(String.self, forKey: .code),
-                  let amount = try? rewardContainer.decode(Decimal.self, forKey: .amount) else {
+                  let amount = try? rewardContainer.decode(Int.self, forKey: .amount) else {
                 Logger.warn(
                     Strings.backendError.malformed_reward_verification_reward_payload(type: rewardType)
                 )
